+++ src/main/java/net/spell_engine/api/spell/ParticleBatch.java
@@ -35,12 +35,15 @@ public class ParticleBatch { public ParticleBatch() { }
     public float min_speed = 0;
     public float max_speed = 1;
     public float angle = 0;
+    // Static position offset
     public float extent = 0;
+    // Motion based position offset
+    public float pre_spawn_travel = 0;
     public boolean invert = false;
 
     public ParticleBatch(
             String particle_id, Shape shape, Origin origin, Rotation rotation,
-            float count, float min_speed, float max_speed, float angle, float extent, boolean invert) {
+            float count, float min_speed, float max_speed, float angle, float extent, float pre_spawn_travel, boolean invert) {
         this.particle_id = particle_id;
         this.shape = shape;
         this.origin = origin;
@@ -50,19 +53,23 @@ public class ParticleBatch { public ParticleBatch() { }
         this.max_speed = max_speed;
         this.angle = angle;
         this.extent = extent;
+        this.pre_spawn_travel = pre_spawn_travel;
+        this.invert = invert;
     }
 
     // Compatibility constructors
 
+    @Deprecated
     public ParticleBatch(String particle_id, Shape shape, Origin origin, Rotation rotation,
                          float count, float min_speed, float max_speed, float angle) {
         this(particle_id, shape, origin, rotation, count, min_speed, max_speed, angle, 0);
     }
 
+    @Deprecated
     public ParticleBatch(
             String particle_id, Shape shape, Origin origin, Rotation rotation,
             float count, float min_speed, float max_speed, float angle, float extent) {
-        this(particle_id, shape, origin, rotation, count, min_speed, max_speed, angle, extent, false);
+        this(particle_id, shape, origin, rotation, count, min_speed, max_speed, angle, extent, 0, false);
     }
 
     // Copy
@@ -77,6 +84,7 @@ public class ParticleBatch { public ParticleBatch() { }
             other.max_speed,
             other.angle,
             other.extent,
+            other.pre_spawn_travel,
             other.invert);
     }
 }