+++ src/main/java/net/spell_engine/network/Packets.java
@@ -127,6 +127,7 @@ public class Packets {
             buffer.writeFloat(batch.max_speed);
             buffer.writeFloat(batch.angle);
             buffer.writeFloat(batch.extent);
+            buffer.writeFloat(batch.pre_spawn_travel);
             buffer.writeBoolean(batch.invert);
         }
 
@@ -136,6 +137,7 @@ public class Packets {
                     ParticleBatch.Shape.values()[buffer.readInt()],
                     ParticleBatch.Origin.values()[buffer.readInt()],
                     ParticleBatch.Rotation.from(buffer.readInt()),
+                    buffer.readFloat(),
                     buffer.readFloat(),
                     buffer.readFloat(),
                     buffer.readFloat(),